test(codemods): split codemod coverage into focused Vitest fixture suites
